+++ Changelog.md
@@ -3,6 +3,7 @@
 ## 0.2.1 (unreleased)
 
 - Added search and a view for recently modified pages, all in one!
+- Added the ability to create new files via the web interface
 - Goldfish now automatically creates the data directory if it doesn't exist already
 
 ## 0.2.0 (2015-05-15)